feat: navbar, tailwind, routing, logo/icon addition
