+++ frontend/index.html
@@ -2,12 +2,12 @@
 <html>
   <head>
     <meta charset="utf-8" />
-    <title> Fluke • To be or not to be </title>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css">
+    <title>Fluke • Chat Done Right</title>
     <link data-trunk rel="rust" />
     <link data-trunk rel="scss" href="./index.scss" />
-    <link data-trunk rel="copy-file" href="assets/twitter_header_photo_1.png"/>
-    <link data-trunk rel="copy-file" href="assets/favicon.ico" />
+    <link rel="stylesheet" href="/tailwind.css" />
+    <link data-trunk rel="copy-file" href="dist/logo_transparent.png" />
+    <link data-trunk rel="icon" href="assets/favicon.ico" />
   </head>
   <body></body>
 </html>
