+++ frontend/index.scss
@@ -1,170 +1,183 @@
 :root {
-    --text-primary: rgba(255, 255, 255, 0.87);
-    --text-secondary: rgba(255, 255, 255, 0.6);
-    --text-disabled: rgba(255, 255, 255, 0.38);
-    --action-active: rgba(255, 255, 255, 0.54);
-    --action-hover: rgba(255, 255, 255, 0.04);
-    --action-selected: rgba(255, 255, 255, 0.08);
-    --action-disabled: rgba(255, 255, 255, 0.26);
-    --action-disabledBackground: rgba(255, 255, 255, 0.12);
-    --background-default: #121212;
-    --background-paper: #1E1E1E;
-    --divider: rgba(255, 255, 255, 0.12);
+  --text-primary: rgba(255, 255, 255, 0.87);
+  --text-secondary: rgba(255, 255, 255, 0.6);
+  --text-disabled: rgba(255, 255, 255, 0.38);
+  --action-active: rgba(255, 255, 255, 0.54);
+  --action-hover: rgba(255, 255, 255, 0.04);
+  --action-selected: rgba(255, 255, 255, 0.08);
+  --action-disabled: rgba(255, 255, 255, 0.26);
+  --action-disabledBackground: rgba(255, 255, 255, 0.12);
+  --background-default: #121212;
+  --background-paper: #1e1e1e;
+  --divider: rgba(255, 255, 255, 0.12);
 }
 
 html {
-    scroll-behavior: smooth;
-    margin: 0;
-    padding: 0;
-    background: #2d3131;
-    height: 100%;
-  
-    &,
-    body {
-        color: var(--text-primary);
-        background-color: var(--background-default);
-    }
-}
+  scroll-behavior: smooth;
+  margin: 0;
+  padding: 0;
+  height: 100%;
 
-main.home {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-h1, h2, h3, h4, h5, h6,
-p, a, span, li {
+  &,
+  body {
     color: var(--text-primary);
+    background-color: var(--background-default);
+    height: 100%;
+  }
 }
 
-p, a, span, li {
-    color: var(--text-secondary);
+main {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+p,
+a,
+span,
+li {
+  color: var(--text-primary);
+}
+
+p,
+a,
+span,
+li {
+  color: var(--text-secondary);
 }
 
 .disabled {
-    color: var(--text-disabled);
+  color: var(--text-disabled);
 }
 
 button {
-    color: var(--action-active);
-    background-color: var(--action-disabledBackground);
-    border-radius: 10px;
+  color: var(--action-active);
+  background-color: var(--action-disabledBackground);
+  border-radius: 10px;
 
-    &:hover {
-        background-color: var(--action-hover);
-    }
-  
-    &:active {
-        background-color: var(--action-selected);
-    }
+  &:hover {
+    background-color: var(--action-hover);
+    cursor: pointer;
+  }
 
-    &:disabled {
-        color: var(--action-disabled);
-    }
+  &:active {
+    background-color: var(--action-selected);
+  }
 
-    &.button-primary {
-        grid-column: 1 / 3; /* This makes the button span the entire width of the grid. */
-        justify-self: center; /* This centers the button within the grid cell. */
-    }
+  &:disabled {
+    color: var(--action-disabled);
+  }
+
+  &.button-primary {
+    grid-column: 1 / 3; /* This makes the button span the entire width of the grid. */
+    justify-self: center; /* This centers the button within the grid cell. */
+  }
 }
 
 .divider {
-    border-color: var(--divider);
+  border-color: var(--divider);
 }
 
-#outer, #content {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100%;
+#outer,
+#content {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
 }
 
 .logo {
-    max-width: 50vw;
+  max-width: 50vw;
 }
 
 #links {
-    width: 100%;
-    text-align: center;
+  width: 100%;
+  text-align: center;
 }
 
 .poem {
-    padding: 2rem;
-    color: #5dacbd;
-    font-family: monospace;
-    font-size: 0.8rem;
-    text-align: center;
-    margin: 0;
-    white-space: pre-wrap;
+  padding: 2rem;
+  color: #5dacbd;
+  font-family: monospace;
+  font-size: 0.8rem;
+  text-align: center;
+  margin: 0;
+  white-space: pre-wrap;
 }
 
 #wrapper {
-    width: 70%;
-    margin: auto;
+  width: 70%;
+  margin: auto;
 }
 
 #title {
-    font-size: 2rem;
-    text-align: center;
+  font-size: 2rem;
+  text-align: center;
 }
 
 .header {
-    background-color: #333;
-    overflow: hidden;
+  background-color: #333;
+  overflow: hidden;
 }
 
 .nav-item {
-    display: inline-block;
-    padding: 10px;
-  
-    &:hover {
-        background-color: #ddd;
-        color: black;
-    }
-  
-    &:active {
-        background-color: #4CAF50;
-        color: white;
-    }
+  display: inline-block;
+  padding: 10px;
+
+  &:hover {
+    background-color: #ddd;
+    color: black;
+  }
+
+  &:active {
+    background-color: #4caf50;
+    color: white;
+  }
 }
 
 .nav-link {
-    color: #f2f2f2;
-    text-align: center;
-    text-decoration: none;
-    font-size: 17px;
+  color: #f2f2f2;
+  text-align: center;
+  text-decoration: none;
+  font-size: 17px;
 }
 
 .registration-form {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 10px;
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: #232323;
-    border-radius: 10px;
-    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 10px;
+  max-width: 600px;
+  margin: 0 auto;
+  padding: 20px;
+  background-color: #232323;
+  border-radius: 10px;
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
 }
 
 label {
-    padding: 10px;
-    color: #ffffff;
+  padding: 10px;
+  color: #ffffff;
 }
 
-input[type="text"], 
-input[type="email"], 
+input[type="text"],
+input[type="email"],
 input[type="password"] {
-    padding: 10px;
-    background-color: #333333;
-    color: #ffffff;
-    border: none;
-    border-radius: 5px;
+  padding: 10px;
+  background-color: #333333;
+  color: #ffffff;
+  border: none;
+  border-radius: 5px;
 }
 
 .error-text {
-    grid-column: span 2;
-    color: #FF0000;
+  grid-column: span 2;
+  color: #ff0000;
 }
