+++ frontend/src/app.rs
@@ -1,11 +1,14 @@
+use crate::views::home::Home;
+use crate::views::signup::SignupForm;
 use yew::prelude::*;
 use yew_router::prelude::*;
-use crate::views::home::Home;
 
 #[derive(Clone, PartialEq, Routable)]
 pub enum Route {
     #[at("/")]
     Home,
+    #[at("/login")]
+    SignupForm,
     #[not_found]
     #[at("/404")]
     NotFound,
@@ -14,6 +17,7 @@ pub enum Route {
 pub fn switch(routes: Route) -> Html {
     match routes {
         Route::Home => html! { <Home /> },
+        Route::SignupForm => html! { <SignupForm /> },
         Route::NotFound => html! { <h1>{ "404" }</h1> },
     }
 }
