+++ frontend/src/components/mod.rs
@@ -1,3 +1,3 @@
-pub mod header;
+pub mod button;
 pub mod input;
-pub mod landing;
+pub mod navbar;
