+++ frontend/src/views/mod.rs
@@ -1,1 +1,3 @@
+pub mod about;
 pub mod home;
+pub mod signup;
